Style sign up and footer section
Applied the final styling to the sign up section and footer.
Updated the HTML again by moving elements and adding some extra
divs with classes to allow for easier alignment and customization.
Will periodically return to the application to refactor any of the
code that can benefit from refactoring into a much cleaner looking
website.